requirements.txt
requests
pytest
pytest-html
pytest-order
allure-pytest
alchemize
selenium
python-dotenv